requirements.txt
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.3
ibm-watsonx-ai==1.2.1
elasticsearch==8.12.0
pytesseract==0.3.13
pillow==11.1.0
botocore==1.35.15
ibm-cos-sdk==2.12.0